Refactor : 디자인 프롭 빼둠
