+++ src/pages/mission/MissionImage.tsx
@@ -1,6 +1,7 @@
 import door_clover1 from '@/assets/mission/door_clover1.png';
 import door_clover2 from '@/assets/mission/door_clover2.png';
 import door_clover3 from '@/assets/mission/door_clover3.png';
+import * as style from './mission.style';
 
 interface IMissionImageProps {
     clearedCount: number;
@@ -13,10 +14,7 @@ const MissionImage = ({ clearedCount }: IMissionImageProps) => {
         <img
             src={imgList[clearedCount - 1]}
             alt='클로버 미션 이미지'
-            style={{
-                width: '100%',
-                aspectRatio: '22.5 / 16.25',
-            }}
+            style={style.missionImageStyle}
         />
     );
 };
