+++ src/pages/mission/missionCard/IconBadge.tsx
@@ -1,6 +1,6 @@
 import { Avatar, Typography } from '@mui/material';
 import Clover from '@/assets/clover.png';
-import * as styles from './Mission.style';
+import * as styles from './MissionCard.style';
 
 export interface IIconBadgeProps {
     type: 'clover' | 'my' | 'number';
